BUGFIX: Add default formats to monocle routes
This fixes an annoying bug, through which flow couldn't figure out the correct route to redirect to after login. So, when your session timed out, you got to see the Neos login screen after which you got redirected to a 404 page.

By adding `@format` to the default values, this redirect works now.