BUGFIX: Handle routePath detection behind browsersync-proxy
The requests sent by browser sync have not protocol attached. This makes flow create urls without protocol aswell but the href
that is used by the browser does still contains the url. That makes the detection wether or not an additional routePath was given unstable.

This change fixes this by comparing the pathname instead of href.