BUGFIX: Render prototypes and props in isolation by using a wrapper component
This makes sure the styleguide props are only evaluated for the currently rendered prototype
and that the fusion of the styleguide props does not interfere with the default fusion.

Now all props have to be passed explicitly to the child-components as styleguide props are only preset
for the actually rendered component.