FEATURE: Allow defining of whole props and propSets as Fusion prototypes or expressions
Defining whole props and propSets as Expressions or Fusion prototypes allows to
centralize the fixtures and even extract them into separate json files.

When props or propSets cannot be evaluated directly the keys below are evaluated
as RawArray as before so this change is a non breaking feature.

```
prototype (Vendor.Site:Example) < prototype(Neos.Fusion:Component) {
    @styleguide {
        propSets {
	    fromEelExpression = ${{title: 'hello world'}}
	    fromFusionPrototype = Vendor.Site:Example.Fixtures
	    fromJsonFile = ${Json.parse(File.readFile('_path_to_fixctures.json'), true)}	}
    }
}
```