FEATURE: Prototypes can define `@styleguide.container` for preview
Often components have to be rendered in the styleguide inside another component. In this case a `container`
can be defined in the styleguide annotation. The container is the applied as processor to the uppermost prototype.

```
prototype(Vendor.Site:ExampleComponent) < prototype(Neos.Fusion:Component) {
    @styleguide {
        container = Vendor.Site:ExampleContainer
    }

    renderer = afx`
        <h1>Hello World</h1>
    `
}
```

The `container` prototype has to accept the prop `content` that will contain the rendered prototype.

```
prototype(Vendor.Site:ExampleContainer) < prototype(Neos.Fusion:Component) {
    content = null
    renderer = afx`
        <div class="container">{props.content}</div>
    `
}
```

*Please note that containers are only rendered for the uppermost prototype that is rendered in the styleguide. For all other
prototype only the renderer is evaluated. This also applies to prototypes rendered by `Sitegeist.Monocle:Preview.Prototype`.*