FEATURE: Smarter and configurable determination of defaultPackageKey
If a domain can be identified from the request monocle will use this to determine the initial sitePackage.
If that does not work a newly added setting `defaultPackageKey` is used. If this setting was not set the key of
the first site package or the first key in the `packages` configuration is used.